chore: fix require cycles
