+++ example/src/Examples/ActivityIndicatorExample.tsx
@@ -9,7 +9,7 @@ import {
   MD3Colors,
 } from 'react-native-paper';
 
-import { useExampleTheme } from '..';
+import { useExampleTheme } from '../hooks/useExampleTheme';
 import ScreenWrapper from '../ScreenWrapper';
 
 const ActivityIndicatorExample = () => {
