+++ example/src/Examples/BadgeExample.tsx
@@ -11,7 +11,7 @@ import {
   Switch,
 } from 'react-native-paper';
 
-import { useExampleTheme } from '..';
+import { useExampleTheme } from '../hooks/useExampleTheme';
 import ScreenWrapper from '../ScreenWrapper';
 
 const BadgeExample = () => {
