+++ example/src/Examples/ChipExample.tsx
@@ -11,7 +11,7 @@ import {
   Text,
 } from 'react-native-paper';
 
-import { useExampleTheme } from '..';
+import { useExampleTheme } from '../hooks/useExampleTheme';
 import ScreenWrapper from '../ScreenWrapper';
 
 const ChipExample = () => {
