+++ example/src/Examples/IconButtonExample.tsx
@@ -3,7 +3,7 @@ import { StyleSheet, View } from 'react-native';
 
 import { IconButton, List, MD2Colors, MD3Colors } from 'react-native-paper';
 
-import { useExampleTheme } from '..';
+import { useExampleTheme } from '../hooks/useExampleTheme';
 import ScreenWrapper from '../ScreenWrapper';
 
 const ButtonExample = () => {
