+++ example/src/Examples/ProgressBarExample.tsx
@@ -10,7 +10,7 @@ import {
   Text,
 } from 'react-native-paper';
 
-import { useExampleTheme } from '..';
+import { useExampleTheme } from '../hooks/useExampleTheme';
 import ScreenWrapper from '../ScreenWrapper';
 
 class ClassProgressBar extends React.Component {
