+++ example/src/Examples/SearchbarExample.tsx
@@ -11,7 +11,7 @@ import {
   Avatar,
 } from 'react-native-paper';
 
-import { useExampleTheme } from '..';
+import { useExampleTheme } from '../hooks/useExampleTheme';
 import ScreenWrapper from '../ScreenWrapper';
 
 type Props = {
