+++ example/src/Examples/SurfaceExample.tsx
@@ -9,7 +9,7 @@ import {
   List,
 } from 'react-native-paper';
 
-import { useExampleTheme } from '..';
+import { useExampleTheme } from '../hooks/useExampleTheme';
 import ScreenWrapper from '../ScreenWrapper';
 
 const SurfaceExample = () => {
