+++ example/src/Examples/SwitchExample.tsx
@@ -10,7 +10,7 @@ import {
   TouchableRipple,
 } from 'react-native-paper';
 
-import { useExampleTheme } from '..';
+import { useExampleTheme } from '../hooks/useExampleTheme';
 import ScreenWrapper from '../ScreenWrapper';
 
 const SwitchExample = () => {
