+++ example/src/hooks/useExampleTheme.ts
@@ -1,0 +1,3 @@
+import { MD2Theme, MD3Theme, useTheme } from 'react-native-paper';
+
+export const useExampleTheme = () => useTheme<MD2Theme | MD3Theme>();
